reorganize to use builder
