+++ cipher_reactivity/cipher_reactivity/__init__.py
@@ -1,3 +1,4 @@
+from . import builder
 from .client import AskcosClient
 from .sascorer import SAScorer
 from .trigger import synthesis_trigger
